modified:   task1.exe 	modified:   task10.cpp 	modified:   task10.exe 	modified:   task11.exe 	modified:   task12.exe 	modified:   task2.exe 	modified:   task3.exe 	modified:   task4.cpp 	modified:   task4.exe 	modified:   task5.exe 	modified:   task6.exe 	modified:   task8.exe 	modified:   task9.exe

+++ task10.cpp
@@ -11,7 +11,7 @@ void reverse(string &str, int count = 100) {
     for (int i = count - 1; i >= 0; i--) { 
         str += temp[i]; 
     }
-    for (int i = count; i < temp.length(); i++){
+    for (int i = count; i < (int)temp.length(); i++){
         str += temp[i]; 
     }
 }
